+++ Editor/Abstraction/EditorUtility.cs
@@ -34,5 +34,8 @@ namespace Willykc.Templ.Editor.Abstraction
 
         void IEditorUtility.DisplayProgressBar(string title, string info, float progress) =>
             UnityEditor.EditorUtility.DisplayProgressBar(title, info, progress);
+
+        bool IEditorUtility.DisplayDialog(string title, string message, string ok) =>
+            UnityEditor.EditorUtility.DisplayDialog(title, message, ok);
     }
 }